requirements.txt
aiodns==2.0.0
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
brotlipy==0.7.0
cchardet==2.1.7
cffi==1.14.3
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
psutil==5.7.3
pycares==3.1.1
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.15.0
yarl==1.5.1
